Move helm-specific helper functions to a separate itest/helm.go file.
